Move stuff to the client side properly
